requirements.txt
requests>=2.28.0
google-cloud-bigquery>=3.4.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0 